tests: Consolidate ontology validators into single file per QA directive
This commit resolves a regression identified in the QA review:
1. Deleted the phantom file `tests/contracts/test_ontology_validators_additional.py` which violated the "No File Sprawl" mandate.
2. Restored `tests/contracts/test_ontology_validators.py` to contain the unified validator tests.
3. Restored the `_cached_hash` tests inside `test_coreason_base_state_hash` that had been accidentally removed.
4. Maintained the deterministic sorting tests for `ComputeEngineProfile`, `PermissionBoundaryPolicy`, `ActivationSteeringContract`, `AdjudicationRubricProfile`, and `RedactionPolicy`.
5. Adjusted `DynamicLayoutManifest` syntax error test to assert that `SyntaxError`s are silently handled, allowing the literal layout strings without throwing an execution error.

All tests, formatting, typing, and linters pass cleanly.
